requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
pydantic==1.10.10
python-dotenv==1.0.0
SQLAlchemy==1.4.49
alembic==1.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.1
boto3==1.28.76
requests==2.31.0
email-validator==1.2.0
redis>=4.6.0
piper-tts>=1.2.0
numpy>=1.20.0